requirements_backup.txt
# Web Framework
fastapi
uvicorn[standard]

# HTTP Client for OpenRouter API
httpx

# Data Validation and Settings
pydantic
pydantic-settings

# Image Processing and OCR
Pillow

# Utilities
python-multipart
python-dotenv

google-generativeai
groq

pandas
numpy

langchain

langchain-google-genai

# pip install pip-tools
# pip-compile --generate-hashes requirements_backup.txt
